Move provider and tfgen into provider directory
This commit prepares us for declaring a Go Module for the SDK. All of
the pieces associated with the provider now live in the provider/
subdirectory, and the formerly root go.mod is now in effect only for the
provider itself.

Code generation updating to the latest version of each dependency is in
the following commit.