Change provider token parameter to be optional
Currently, the provider made the token to be required. The setup
of the client within the provider allows the use of the Vault EnvVars
that would allow vault to work locally. So the token itself is not
100% required as there is a default. We do use the token that the user
specifies as the token if it is specified, otherwise we default to
the token that the client creates

This also means a user doesn't need to pass a token and therefore,
it won't be stored in the state