Add alias for AuthBackendRoleSecretId -> AuthBackendRoleSecretID (#202)
* Add alias for AuthBackendRoleSecretId -> AuthBackendRoleSecretID

Fixes #201.

In https://github.com/pulumi/pulumi-vault/pull/190 a case change for the existing
type token leaked through. This change adds an alias to ensure existing resources
with the old type token can be deduped. Note that a code change might still be required
but that is in essence not seen as a breaking change (an errant update due to a missing
alias is breaking however, which this fixes). FWIW this provider mixes conventions
for type tokens. We have resorted to using the current convention for new type tokens.
We don't intend to update tokens for existing resources however.

* Actually add the SDKs