Use version 0.0.9 of pulumi-package-publisher (#322)
The version of pulumi-package-publisher used in the latest release does
not account for the old nodejs install scripts, that the version of the
bridge we are using still generates. Version 0.0.9 has a hotfix meant to
address an npm install failure that expects these scripts.

We can verify this works after merge to default branch, and then
installing `latest`. Then we can push a patch.

Read context here: https://github.com/pulumi/pulumi-github/pull/377

Fixes #321.